CSS work

+++ final_project/client/src/components/Navbar.js
@@ -1,4 +1,5 @@
-import { Link, useNavigate } from "react-router-dom";
+import { Link, useNavigate, useLocation } from "react-router-dom";
+import { useEffect, useState } from "react";
 import Button from "@mui/material/Button";
 import Stack from "@mui/material/Stack";
 import axios from "axios";
@@ -6,10 +7,19 @@ import React, { useContext } from "react";
 import { AppContext } from "../App";
 
 const Navbar = () => {
+  const location = useLocation();
+  const [paths, setPaths] = useState([]);
   const { setAccessToken } = useContext(AppContext);
   const navigate = useNavigate();
 
-  const logout = async () => {
+  useEffect(() => {
+    if (location.pathname !== "/login" && location.pathname !== "/register")
+      setPaths(["game", "leaderboard", "logout"]);
+    else setPaths(["login", "register"]);
+  }, [location]);
+
+  const logout = async (e) => {
+    if (e.target.id !== "logout") return;
     try {
       const response = await axios.delete(
         "/logout",
@@ -33,22 +43,19 @@ const Navbar = () => {
   return (
     <div>
       <Stack spacing={2} direction="row" position="fixed">
-        <Button component={Link} to={"/login"}>
-          Login
-        </Button>
-        <Button component={Link} to={"/register"}>
-          Register
-        </Button>
-        <Button component={Link} to={"/"}>
-          Home
-        </Button>
-        <Button component={Link} to={"/game"}>
-          Tetris
-        </Button>
-        <Button component={Link} to={"/leaderboard"}>
-          Leaderboard
-        </Button>
-        <Button onClick={logout}>Logout</Button>
+        {paths.map((path, index) => {
+          return (
+            <Button
+              component={Link}
+              to={`/${path}`}
+              key={index}
+              onClick={logout}
+              id={path}
+            >
+              {path}
+            </Button>
+          );
+        })}
       </Stack>
     </div>
   );
